refactor: move callout label/value helpers from common to core
These helpers are only consumed by server-side CSV ingestion today. Per
the monorepo convention, backend-only utilities live in @beabee/core, not
@beabee/beabee-common (which is for code shared between backend and
frontend/client). Tests moved alongside.
