+++ apps/backend-cli/src/actions/database/import-callout-responses.ts
@@ -20,13 +20,15 @@ import {
   CalloutResponseAnswerFileUpload,
   CalloutResponseAnswersSlide,
   getCalloutComponents,
-  getSelectableValueFromLabel,
-  parseSelectableMultiValue,
 } from '@beabee/beabee-common';
 import { config } from '@beabee/core/config';
 import { getRepository, runTransaction } from '@beabee/core/database';
 import { Callout, CalloutResponse, Contact } from '@beabee/core/models';
 import { runApp } from '@beabee/core/server';
+import {
+  getSelectableValueFromLabel,
+  parseSelectableMultiValue,
+} from '@beabee/core/utils';
 
 import { isURL } from 'class-validator';
 import { parse } from 'csv-parse';
