+++ packages/core/src/utils/index.ts
@@ -1,4 +1,5 @@
 // TODO: Fix circular dependency problem to export all utils
+export * from './callouts';
 export * from './file';
 export * from './email';
 export * from './rate-limiter';
